Created server with JWT issuing and verification.
Creation and deletion of users are produced to Kafka
Task service reads from kafka and updates its DB.
Task service creates events for create, reassign and close task.
Reassign supposed to work with outbox: all not-closed tasks are put into the separate table, job reads it, assign to a user and deletes.

Not done:
 - Auth is not working yet, request filtering TBD
 - Job, that reads from reassign outbox (reassign of single task is in place)
 - A little bit messy with data fields, to be cleaned